+++ src/rules/validation/abstract/base/RuleBlacklistBase.sol
@@ -98,6 +98,7 @@ abstract contract RuleBlacklistBase is RuleAddressSet, RuleNFTAdapter, RuleBlack
         // Advertise IAddressList: this rule manages an address set and is callable through
         // the IAddressList interface.
         return interfaceId == AddressListInterfaceId.IADDRESS_LIST_INTERFACE_ID
+            || interfaceId == AddressListInterfaceId.IADDRESS_LIST_BATCH_QUERY_INTERFACE_ID
             || RuleTransferValidation.supportsInterface(interfaceId);
     }